**Subject:**
feat: Add icon to the project and attribute its author in README

**Body:**

This commit adds an icon (`icon.png`) to the project, as well as credits
the icon's creator in the `README.md` file.

Changes include:

1. Added a new binary file `icon.png` to the project.
2. Updated `package.json` to include a declaration of project's icon file by
   adding `"icon": "icon.png"` to the file.
3. Updated `README.md` by appending an attribution to the icon's original creator -
   `<a href="https://www.flaticon.com/free-icons/generator" title="generator icons">Generator icons created by edt.im - Flaticon </a>`

Adding an icon improves the visual element in the README and respects the
author's right by attributing the icon properly.